Move optimizer traits to session crate
Update query planning APIs to accept the Session trait and move physical optimizer traits into datafusion-session while preserving re-exports from datafusion-physical-optimizer. Document the new trait location in the upgrade guide.

AI Disclosure: This code was written in part by an AI agent.

(cherry picked from commit 6ef2d3cfdd843594861a2e05234c1a81d0e31eec)
