+++ datafusion-examples/examples/relation_planner/table_sample.rs
@@ -103,8 +103,8 @@ use tonic::async_trait;
 use datafusion::optimizer::simplify_expressions::simplify_literal::parse_literal;
 use datafusion::{
     execution::{
-        RecordBatchStream, SendableRecordBatchStream, SessionState, SessionStateBuilder,
-        TaskContext, context::QueryPlanner,
+        RecordBatchStream, SendableRecordBatchStream, Session, SessionState,
+        SessionStateBuilder, TaskContext, context::QueryPlanner,
     },
     physical_expr::EquivalenceProperties,
     physical_plan::{
@@ -563,8 +563,11 @@ impl QueryPlanner for TableSampleQueryPlanner {
     async fn create_physical_plan(
         &self,
         logical_plan: &LogicalPlan,
-        session_state: &SessionState,
+        session: &dyn Session,
     ) -> Result<Arc<dyn ExecutionPlan>> {
+        let Some(session_state) = session.as_any().downcast_ref::<SessionState>() else {
+            return internal_err!("TableSampleQueryPlanner requires a SessionState");
+        };
         let planner = DefaultPhysicalPlanner::with_extension_planners(vec![Arc::new(
             TableSampleExtensionPlanner,
         )]);
