+++ datafusion/core/src/execution/mod.rs
@@ -28,3 +28,4 @@ pub use session_state_defaults::SessionStateDefaults;
 // backwards compatibility
 pub use crate::datasource::file_format::options;
 pub use datafusion_execution::*;
+pub use datafusion_session::Session;
