+++ datafusion/physical-optimizer/src/optimizer.rs
@@ -17,7 +17,6 @@
 
 //! Physical optimizer traits
 
-use std::fmt::Debug;
 use std::sync::Arc;
 
 use crate::aggregate_statistics::AggregateStatistics;
@@ -40,29 +39,8 @@ use crate::hash_join_buffering::HashJoinBuffering;
 use crate::limit_pushdown_past_window::LimitPushPastWindows;
 use crate::pushdown_sort::PushdownSort;
 use crate::window_topn::WindowTopN;
-use datafusion_common::Result;
 use datafusion_common::config::ConfigOptions;
-use datafusion_physical_plan::ExecutionPlan;
-use datafusion_physical_plan::operator_statistics::StatisticsRegistry;
-
-/// Context available to physical optimizer rules.
-///
-/// This trait provides access to configuration options and optional statistics
-/// registry for enhanced statistics lookup. It allows optimizer rules to access
-/// extended context without changing the core [`PhysicalOptimizerRule::optimize`]
-/// signature.
-pub trait PhysicalOptimizerContext: Send + Sync {
-    /// Returns the configuration options.
-    fn config_options(&self) -> &ConfigOptions;
-
-    /// Returns the statistics registry for enhanced statistics lookup.
-    ///
-    /// Returns `None` if no registry is configured, in which case rules
-    /// should fall back to using `ExecutionPlan::partition_statistics()`.
-    fn statistics_registry(&self) -> Option<&StatisticsRegistry> {
-        None
-    }
-}
+use datafusion_session::{PhysicalOptimizerContext, PhysicalOptimizerRule};
 
 /// Simple context wrapping [`ConfigOptions`] for backward compatibility.
 ///
@@ -84,47 +62,6 @@ impl PhysicalOptimizerContext for ConfigOnlyContext<'_> {
     fn config_options(&self) -> &ConfigOptions {
         self.config
     }
-}
-
-/// `PhysicalOptimizerRule` transforms one ['ExecutionPlan'] into another which
-/// computes the same results, but in a potentially more efficient way.
-///
-/// Use [`SessionState::add_physical_optimizer_rule`] to register additional
-/// `PhysicalOptimizerRule`s.
-///
-/// [`SessionState::add_physical_optimizer_rule`]: https://docs.rs/datafusion/latest/datafusion/execution/session_state/struct.SessionState.html#method.add_physical_optimizer_rule
-pub trait PhysicalOptimizerRule: Debug + std::any::Any {
-    /// Rewrite `plan` to an optimized form.
-    ///
-    /// This is the primary optimization method. For rules that need access to
-    /// the statistics registry, override [`optimize_with_context`](Self::optimize_with_context) instead.
-    fn optimize(
-        &self,
-        plan: Arc<dyn ExecutionPlan>,
-        config: &ConfigOptions,
-    ) -> Result<Arc<dyn ExecutionPlan>>;
-
-    /// Rewrite `plan` with access to extended context (statistics registry, etc.).
-    ///
-    /// Override this method if you need access to the statistics registry for
-    /// enhanced statistics lookup. The default implementation simply calls
-    /// [`optimize`](Self::optimize) with the config options from the context.
-    fn optimize_with_context(
-        &self,
-        plan: Arc<dyn ExecutionPlan>,
-        context: &dyn PhysicalOptimizerContext,
-    ) -> Result<Arc<dyn ExecutionPlan>> {
-        self.optimize(plan, context.config_options())
-    }
-
-    /// A human readable name for this optimizer rule
-    fn name(&self) -> &str;
-
-    /// A flag to indicate whether the physical planner should validate that the rule will not
-    /// change the schema of the plan after the rewriting.
-    /// Some of the optimization rules might change the nullable properties of the schema
-    /// and should disable the schema check.
-    fn schema_check(&self) -> bool;
 }
 
 /// A rule-based physical optimizer.
