+++ datafusion/session/src/lib.rs
@@ -36,6 +36,8 @@
 //! * Runtime environment configuration
 //! * Query state persistence
 
+pub mod physical_optimizer;
 pub mod session;
 
+pub use crate::physical_optimizer::{PhysicalOptimizerContext, PhysicalOptimizerRule};
 pub use crate::session::{Session, SessionStore};
